+++ src/mobile-token/use-token-toggle-details.ts
