+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_ProfileStack/Profile_TravelTokenScreen/Profile_TravelTokenScreen.tsx
@@ -10,7 +10,7 @@ import {FaqSection} from '@atb/stacks-hierarchy/Root_TabNavigatorStack/TabNav_Pr
 import {ChangeTokenAction} from '@atb/stacks-hierarchy/Root_TabNavigatorStack/TabNav_ProfileStack/Profile_TravelTokenScreen/ChangeTokenAction';
 import {useMobileTokenContextState} from '@atb/mobile-token';
 import {useIsFocused} from '@react-navigation/native';
-import {useTokenToggleDetails} from '@atb/utils/use-token-toggle-details';
+import {useTokenToggleDetails} from '@atb/mobile-token/use-token-toggle-details';
 import {useRemoteConfig} from '@atb/RemoteConfigContext';
 
 type Props = ProfileScreenProps<'Profile_TravelTokenScreen'>;
